tools: run-tests.bat: ifstest: disable SetPointIoReparseDataInvalidTest
This test succeeds on Server 2012 and fails on Server 2019. Investigation
showed that the FSCTL_SET_REPARSE_POINT input buffer length was 23 instead
of less than REPARSE_DATA_BUFFER_HEADER_SIZE(==8) as ifstest claims. This
suggests that WinFsp is not the problem here, but perhaps some OS changes
between Server 2012 and Server 2019. NOTE that we are still using the
ifstest from Server 2012 HCK, which may account for the difference.